+++ nvim/.config/nvim/lua/plugins/todo-comments.lua
@@ -1,13 +1,12 @@
 return {
   "folke/todo-comments.nvim",
   dependencies = { "nvim-lua/plenary.nvim" },
-  event = { "BufReadPost", "BufNewFile" },
-  config = function()
-    require("todo-comments").setup()
-
-    vim.keymap.set("n", "]t", function() require("todo-comments").jump_next() end, { desc = "Next TODO comment" })
-    vim.keymap.set("n", "[t", function() require("todo-comments").jump_prev() end, { desc = "Previous TODO comment" })
-    vim.keymap.set("n", "<leader>ft", "<cmd>TodoTelescope<cr>", { desc = "Find TODOs" })
-    vim.keymap.set("n", "<leader>xt", "<cmd>Trouble todo toggle<cr>", { desc = "TODOs (Trouble)" })
-  end,
+  event = "VeryLazy",
+  keys = {
+    { "]t", function() require("todo-comments").jump_next() end, desc = "Next TODO comment" },
+    { "[t", function() require("todo-comments").jump_prev() end, desc = "Previous TODO comment" },
+    { "<leader>ft", "<cmd>TodoTelescope<cr>", desc = "Find TODOs" },
+    { "<leader>xt", "<cmd>Trouble todo toggle<cr>", desc = "TODOs (Trouble)" },
+  },
+  opts = {},
 }